pkg/asset: Switch from github.com -> gopkg.in for survey
This is the recommended import URI [1].  And survey's internal imports
use the gopkg.in form, so we're already vendoring it; this change will
let us drop the redundant github.com copies.

The explicit package name is not strictly necessary, but I like to set
it when the final segment of the import URL does not match the package
name declared within the imported package.

[1]: https://github.com/AlecAivazis/survey/tree/v1.6.2#versioning