hack/get-terraform: Support gunzip as an alternative
Michael had gunzip installed, but not funzip [1], and while funzip is
specifically written to extract the first file from a ZIP archive,
gunzip seems to be able to do the same thing (although I haven't found
docs to back this up).

This commit addss a helper function 'have' to get a more compact
'command' call.  And it adds a new FUNZIP variable (which the user can
override if they want, using the ${parameter:-word} syntax [2])
defaulting to 'funzip'.  If the initial command is missing, we fall
back to gunzip if it's available, and finally exit with an error if it
wasn't.

There's a race between the initial "do we have ${FUNZIP}?" check and
the final 'command -V' [3] call used to generate the "not found" error
message.  But I don't expect many devs to supply the ${FUNZIP} command
during the brief window between checks.

[1]: https://github.com/openshift/installer/pull/356#discussion_r221354378
[2]: http://pubs.opengroup.org/onlinepubs/9699919799/utilities/V3_chap02.html#tag_18_06_02
[3]: http://pubs.opengroup.org/onlinepubs/9699919799/utilities/command.html