data/data/*/inputs: Drop terraform_remote_state
We used to use this for passing asset state from the assets step to
the infra step.  But the new installer performs all asset generation
in Go with a single Terraform step (e.g. see df14588e, *: Drop
Terraform steps, 2018-09-27, #362).  So we'll alwayshave
var.ignition_bootstrap set and we can just use it directly.