cmd/openshift-install: Add Terraform version to 'version' output
And print the error information if we can't extract it.

The .Stderr stuffing is because Terraform just prints usage
information to stdout (not stderr!) if you call it with an
unrecognized command:

  $ terraform does-not-exit >stdout 2>stderr
  $ head -n1 stdout
  Usage: terraform [-version] [-help] <command> [args]
  $ ls -l stderr
  -rw-r--r--. 1 trking trking 0 Oct  2 17:14 stderr

and in this case we want to see that usage information.  Ideally
Terraform would have printed "unrecognized command 'does-no-exist'" or
some such to stderr, but that's a bit bigger than we can work around
here.