README: Add console and cleanup information
This restores the console docs which we'd removed from the README in
feb41e9d (docs: rework documentation, 2018-09-24, #328).  And it moves
the kubeconfig location information over from the libvirt-specific
docs.  Launching the cluster is nice, but these other operations are
important too ;).  Putting them in the README makes increases their
visibility.  It also lets us drom them from the libvirt-specific docs,
now that the libvirt docs link to the README quick-start for these
generic operations.