pkg/rhcos/ami: Start supporting RHCOS channels
openshift/os@76b3ae5d (aws-test: Add a rhcos_test=smoketested tag,
2018-09-25, openshift/os#314) started adding rhcos_tag=alpha tags to
tested AMIs.  The likely plan for RHCOS tagging is [1], although they
don't have stable docs for that yet.  With this commit, we allow the
user to specify whichever channel they want, and we search for that
tag.  If they give us nonsense, we'll now hit AWS, see no images in
the nonsense channel, and report that back to the caller.

Users that don't care about release channels and just want to use the
latest image regardless of channel, can set 'channel' to an empty
string.

[1]: https://github.com/openshift/os/issues/201#issuecomment-423214423