data/libvirt: Resize to 16 GiB
This is the size of the current RHCOS images:

  $ qemu-img info ~/.cache/openshift-install/libvirt/image/2454fcdc0a61a9005fbf916b54fbe332
  image: /home/trking/.cache/openshift-install/libvirt/image/2454fcdc0a61a9005fbf916b54fbe332
  file format: qcow2
  virtual size: 16G (17179869184 bytes)
  disk size: 1.6G
  cluster_size: 65536
  Format specific information:
      compat: 1.1
      lazy refcounts: false
      refcount bits: 16
      corrupt: false

Resizing here will allow the OS folks to stop guessing which size we
need [1,2].  And when we need future resizes, we can update here
instead of bugging the OS team.

Also inline the volume in the libvirt Terraform module.  The
libvirt/volume module indirection was unnecessary indirection.

[1]: https://github.com/openshift/installer/issues/126
[2]: https://github.com/openshift/os/pull/228