assets/bootkube: add etcd service and endpoints for apiservers.
The ip address for the etcd service endpoints are dummy addresses that
should not be used.

The ip address has been chosen from https://tools.ietf.org/html/rfc5737
```
The blocks 192.0.2.0/24 (TEST-NET-1), 198.51.100.0/24 (TEST-NET-2),
and 203.0.113.0/24 (TEST-NET-3) are provided for use in documentation.
```
so that they are don't overlap with any allocated ip addrs.