asset/ignition: clean up bootstrap content
Adding a line break before and after the closing and opening grave ticks
makes this a little more readable and is the style that the rest of the
codebase uses. This does result in an empty line at the top and bottom
of the file, which is why this style cannot be applied to the shell
scripts.