pkg/asset: Expose fields of Asset objects for marshaling into state file
Expose the fields of all Asset/WritableAsset objects so that we can Marshal/Unmarshal them from the state file.
Some of the fields clash with a function name implemented for the struct, so we have to change the field names themselves e.g. files cannot just be Files because there is a Files() function expected by the Asset interface.

Not all fields need be exported actually, because some fields can be constructed from the 'Files' field anyway. This optimization has not been done in this commit, it would perhaps need a custom UnmarshalJSON function to do that.

With this commit, all assets can be saved and loaded back properly. So, an example run of the openshift-install binary will look like this:

$ openshift-install manifests --dir=test --log-level=debug
$ # the above command generates the folders test/{manifests,tectonic} with all the files within
$ # it also genrates the state file in test/.openshift_install_state.json
$ # let's remove the manifest files and generate them again..
$ rm -rf test/{manifests,tectonic}
$ openshift-install manifests --dir=test --log-level=debug
$ # check that everything has been restored without actually generating anything
$ # run the install-config command even to create the install-config.yaml from state file:
$ openshift-install install-config --dir=test --log-level=debug