asset/ignition: add event on bootstrap completion
In order to allow the installer to eventually track the installation
progress, this adds a reporting service which inserts an event when the
bootstrap process has completed. This event can be used to determine
when it is safe to remove the bootstrap node, since it signifies that
the bootstrap control plane has pivoted to the scheduled control plane
and that all of the OpenShift manifests have been inserted.

This no longer enables `tectonic.service` because `progress.service` is
enabled and lists both `tectonic.service` and `bootkube.service` as
dependencies. systemd will start all three services as a result.

In an ideal world, this would make use of sd-notify to signal the
completion of prerequisites but this doesn't work for two reasons.
First, runc doesn't support passing the extra file descriptor used by
sd-notify [1]. Second, systemd doesn't support retrying activations when
a dependency has failed [2]. This change makes use of "breadcrumb" files
to signal completion as a workaround.

In the creation of the event, `date` is invoked and passed a custom
format. This was necessary because Golang requires the `T` separator in
the time package [3] but GNU's `date` uses a space separator. ISO8601
was also tried, but Golang doesn't seem to parse that correctly either
(`date` uses "+0000" to specify timezone but Golang expects "+00:00").

[1]: https://github.com/opencontainers/runc/pull/1807
[2]: https://github.com/systemd/systemd/issues/1312
[3]: https://github.com/golang/go/issues/25937