pkg/asset: TargetableAsset type for special assets that have action associated with them
With the state file supplying asset values instead of them being generated, the cluster 'Asset' is not sufficient to describe that it is a special kind of target.
In this commit, the 'Cluster' WritableAsset is now a TargetableAsset, which has a special meaning - they will have an 'Apply' function that will be executed irrespective of whether 'Generate' is called upon the asset or not.