ignition/bootstrap: add checkpointer and kube-proxy manifests
In order to complete the bootkube start sequence we need to provide
checkpointer pod that has to be observed as "ready" by the bootkube.
Before these manifests were provided by kube-core operator and before
we have the checkpointer operator that provide these, we have to
install them directly temporarely.

The kube-proxy manifests are needed for the tectonic-network operator
in order to talk to the bootkube control plane. These should be owned
by networking team and moved to their operator.