pkg/destroy/bootstrap: Fix cross-filesystem tfstate recovery
Use a copy and then a rename, instead of just a rename.  This avoids
'invalid cross-device link' errors when the source and target
directories are on different filesystems (e.g. because the source is
in a tmpfs under /tmp).

Also fix the terraform.Destroy error handling, and actually return the
wrapped error instead of ignoring it.