data/bootstrap: Pull content out of pkg/asset/ignition/bootstrap
It's easier for humans and linters to find this content if it's not
hidden in Go variables.

Since we're effectively pulling these files from Git now (either at
build time or at run-time depending on release vs. dev mode in
hack/build.sh), I'm being a bit more relaxed about file modes than the
previous implementation.  Files are now either 0555 (if they are in a
'bin' directory) or 0600 (if they aren't).  This is a change for files
like manifests.Manifests, which had previously been 0644.

I've flattened the manifest overrides into a single directly, because
the filenames are sufficient for sorting them by operator.  And all of
the override manifests now have their own comment explaining their
target and eventual location.