asset/ignition: remove bootstrap kubelet debugging
This greatly simplifies the generation of the kubelet service for the
bootstrap node since templating is no longer required. Since the
bootstrap node is no longer a part of the cluster and is temporary, we
should increase the logging verbosity across the board if we find that
it is beneficial. This is preferred over asking our users to customize
their bootstrap node to increase logging.