asset/ignition: properly signal bootstrap complete
3dbbbf6 incorrectly signalled the completion of bootstrap.sh and
tectonic.sh using systemd's ExecStartPost stanza. This doesn't work in
Type=simple services, because systemd runs the ExecStartPost stanza
immediately after spawning the main process. This could be fixed by
changing the service type to "oneshot", where systemd waits for the
ExecStart stanzas to complete, but oneshot services don't respect the
Restart stanza. It's easiest to just signal the completion of these
scripts at the end of the script itself, even if it couples the scripts
to the service unit for progress.service (which is responsible for
checking for the completion of the scripts).