asset/ignition: various bootstrap cleanups
8ffb91c moved the etcd health check from the body of a loop to the
condition of a loop. The `set -e` and `set +e` were required in order to
allow the loop to work (otherwise, the script would have exited on the
first error). Now that the health check is the condition of a loop, the
script won't exit early and these `set` calls are unneeded.

This also adds RemainAfterExit in order to prevent bootkube.service and
tectonic.service from running multiple times. When systemd repeatedly
starts progress.service, it also activates the dependencies, causing the
service to start again if it has exited and deactivated. RemainAfterExit
keeps the services activated so that systemd doesn't restart them.