asset/store: try to reuse assets from state file
Before this change, when the installer found an asset in both the target
directory and the state file, it assumed that the asset should be dirty,
causing the asset and all of its dependants to be regenerated. This is a
small optimization that prevents regenerating dependant assets when the
provided asset is identical to the one in the state file.