rhcos: implement image discovery for new pipeline
RHCOS has moved over to the 2.0 pipeline which has a nicer discovery
mechanism for images. This uses that new pipeline to fetch the latest
QCOW images for the latest build. Unfortunately, there is no "latest"
alias anymore, so the installer has to fetch resources from the Internet
before it can prompt the user. This will go away once the installer
starts pinning to a specific RHCOS release.