cmd/openshift-install/create: Destroy bootstrap on bootstrap-complete
Once the cluster is up, the bootstrap stuff is just a waste of
resources.  With this commit we make a reasonable effort to wait for
the bootstrap-complete event, and, if we see it, we tear down the
bootstrap resources.

The Kubernetes client initialization is based on [1] and the local
wait.go is based on [2].  Both of those are, like this project, under
the Apache 2.0 license.  Once kubernetes/kubernetes#50102 lands, we
can drop the local wait.go and use the upstream implementation.

I've dropped the errors.Cause call from main.go, because it was
dropping the wrapped context, which is useful context for debugging.
That Cause call is from dc118f20 (cmd: switch to already vendored
cobra, 2018-10-08, #429), but that commit doesn't explicitly motivate
the call.

[1]: https://github.com/kubernetes/client-go/blob/v9.0.0/examples/out-of-cluster-client-configuration/main.go
[2]: https://github.com/kubernetes/kubernetes/pull/50102/files