cmd/openshift-install/create: don't log nil errors
The call to cancel isn't a synchronous operation and the closure
executes to completion. As a result, the debug call which logs the error
should be in an else branch so it isn't printed if there is no error.
Without this change, the following would show up in the log:

    INFO API v1.11.0+d4cacc0 up
    DEBUG API not up yet: %!s(<nil>)