asset: trim Store interface and add Destroy to Store
Save and Purge functions don't really seem to belong to Store interface.

Fetch fetches the asset, checkpoints the state file to disk and consumes
all the assets that were loaded from disk to create the asset.

Also fixes the error in save() where it would only save the assets in memory and
drops all the other assets present in the state file but were not fetched into the store's
asset map
eg:
```console
./openshift-install ign configs # state has everything
./openshift-install install-config # state now only has assets uptill install-config, all the ign config assets are lost from statefile

./openshift-install ign configs # state has everything
./openshift-install install-config # state still includes all the state from 'ign-config' run.
```

Destroy deletes the asset from all the possible sources, state file and disk.