asset/ignition: fix bootkube and tectonic retries
Back in 54242f8, RemainAfterExit was added to prevent bootkube.service
and tectonic.service from running multiple times when progress.service
restarted. Due to a systemd bug [1], this causes the restart to be
ignored on those services.

This new approach uses conditions instead. This allows systemd to
activate the services but not actually start them if they have already
completed.

[1]: https://github.com/systemd/systemd/issues/3396