data/aws: Move S3 bucket to bootstrap module
Since 3f4fe574 (cmd/openshift-install: Add 'destroy bootstrap'
command, 2018-10-18, #493), the bootstrap module is targeted for
deletion by 'destroy bootstrap'.  Moving the S3 bucket into the module
means that the bucket, which we only use for hosting the bootstrap
Ignition configuration, will be removed along with the rest of the
bootstrap resources.