vendor: Fork httpcache to support streaming images
Recent RHCOS images are ~1.6 GiB, which is a lot to hold in memory at
once [1].  I've submitted a patch upstream to support streaming [2],
but upstream is pretty quiet, and I won't be surprised if it takes a
while to land there.  In the meantime, I'm pointing the installer at
my personal repository, where I've landed the fix in master.

[1]: https://github.com/openshift/installer/issues/610
[2]: https://github.com/gregjones/httpcache/pull/93