vendor: Fork httpcache to support streaming images
Recent RHCOS images are ~1.6 GiB, which is a lot to hold in memory at
once [1].  I've submitted a patch upstream to support streaming [2],
but upstream rejected it because it introduces an API change.
Reworking the patch to make it acceptable upstream would probably
involve a significant amount of effort.

This patch reduces the memory requirements for the download and unpack
phase from over 20 GiB down to less than 150 MiB, with
rhcos-410.8.20190412.1.

[1]: https://github.com/openshift/installer/issues/610
[2]: https://github.com/gregjones/httpcache/pull/93

Signed-off-by: Stephen Kitt <skitt@redhat.com>