.github/ISSUE_TEMPLATE: Link to the new troubleshooting docs
Make it easier for issue-reporters to find these docs, which will
hopefully help improve issue quality.

Ideally this would be a relative link into the local repository, in
case folks wanted to enable issues in a fork or some such.  But
there's no way (that I've found) to do that which works from both
[1,2].  We can't event use an absolute-path reference
(/openshift/installer/blob/...), because GitHub interprets those as
relative to the filebrowser repo root when rendering [1] but not when
rendering [2]).  Still, while I think hard-coding the org/repo is
ugly, we already do an awful lot of that for our internal Go imports.

[1]: https://github.com/openshift/installer/blob/master/.github/ISSUE_TEMPLATE.md
[2]: https://github.com/openshift/installer/issues/new