pkg/asset/installconfig: Pull AWS region defaults from usual places
The AWS_DEFAULT_REGION environment variable is widely supported,
e.g. by the AWS command line [1], the AWS Go library [2], and
Terraform [3].  The usual AWS configuration also supplies other ways
to declare a default region (e.g. ~/.aws/config [4,5]).  With this
commit, we check those locations instead of assuming us-east-1.  We'll
still prompt the user to confirm.  Personally, I'd have preferred
treating AWS_DEFAULT_REGION as a strong enough choice to not need
confirmation, but Abhinav wanted to prompt for that case too [6].

This change is similar to, but slightly less of a slam-dunk than,
850f7296 (pkg/tfvars: Drop AWS.Profile, 2018-10-15), because I haven't
gotten us out of the region business entirely.  Maybe we'll add
multi-region support or something in the future, and then being
explicit about the regions will be necessary (because there are no
multi-region config conventions that I know of).

[1]: https://docs.aws.amazon.com/cli/latest/userguide/cli-environment.html
[2]: https://godoc.org/github.com/aws/aws-sdk-go/aws/session#hdr-Environment_Variables
[3]: https://www.terraform.io/docs/providers/aws/#environment-variables
[4]: https://docs.aws.amazon.com/cli/latest/userguide/cli-config-files.html
[5]: https://godoc.org/github.com/aws/aws-sdk-go/aws/session#hdr-Creating_Sessions
[6]: https://github.com/openshift/installer/pull/688#discussion_r238968232