terraform: remove references to tectonic from aws terraform modules
The session name when connecting to AWS is changed from starting with TECTONIC_INSTALLER_
to starting with OPENSHIFT_INSTALLER.

The terraform name of the internal route53 zone is changed from terraform_int to
int. The Name tag of the zone is changed from ending with _tectonic_int to ending with _int.

The terraform name of the base route53 zone is changed from tectonic to base.

The terraform name of the api route53 records are changed from tectonic_api_external and
tectonic_api_internal to api_external and api_internal, respectively.

Changes for https://jira.coreos.com/browse/CORS-878