Add LB and DNS (service) vm
We're adding a service VM to support 2, work in progress, features in
the OpenStack platform architecture. This VM runs 2 pods: A dnsmasq one
and an haproxy one.

The dnsmasq pod has been configured to serve the ETCD A and SRV records
on the service VM, forwards requests to Neutron DNS and configures the
bootstrap and master servers to use it.

The haproxy pod has been configured to serve requests for the `api`
hostname that the installer depends on.

This VM is not meant to be put in production but rather allow for
development to progress while the final architecture implementations
for these 2 services are finished.