docs/dev/operators: Drop docs for the CVO approach
That approach should be documented in the CVO itself, since it's not
installer-specific and moving it gets the docs and implementation for
that approach into the same repository.  I've filed [1] to land
dynamic-object docs in the CVO repo (based on some of the content I'm
removing here).  Naming files, etc. are already covered by the
existing CVO documentation.

[1]: https://github.com/openshift/cluster-version-operator/pull/59