aws: Increase default master disk size to 120GB for IO
With colocated masters we are seeing about ~120 IOP/s sustained, and
a 30GB gp2 drive is limited to 100 IOP/s. etcd is ~75% of the write
workload, but we are seeing some syncs take ~1s and occasional
heartbeat latency. Increase master disk by 4x to get slightly more
head room - this would in practice result in about $30-40/month more
disk on top of the ~200/mo the instances cost.