Log more verbosely when waiting for bootstrap completion
First, print the total amount of time we're going to wait.

Now, I had an issue where my cluster was fine but my dev container's
DNS didn't have `127.0.0.1` for my dnsmasq; it would have been a lot more obvious
to me if I clearly saw DNS errors.

Tweak the logging here so that we log when the error message
changes.

Reworked-by: W. Trevor King <wking@tremily.us>